+++ semcore/data-table/__tests__/data-table-virtual-scroll.browser-test.tsx
@@ -9,10 +9,10 @@ test.describe('Vertical Scroll', () => {
     await page.setContent(htmlContent);
 
     await page.keyboard.press('Tab');
-    for (let i = 0; i < 100; i++) {
-      await page.keyboard.press('ArrowDown');
+    for (let i = 0; i < 50; i++) {
+      await page.keyboard.press('ArrowDown', { delay: 50 });
     }
-    await page.waitForSelector('[role="gridcell"][data-ui-name="Body.Cell"][name="id"]:has-text("#100")', { state: 'visible' });
+    await page.waitForSelector('[role="gridcell"][data-ui-name="Body.Cell"][name="id"]:has-text("#50")', { state: 'visible' });
     await expect(page).toHaveScreenshot({ maxDiffPixelRatio: 0.01 });
   });