+++ semcore/typography/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.2.0] - 2025-07-22
+
+### Added
+
+- `formatTags` property for the Text component for enable tags formatting's inside a Text component.
+
 ## [16.1.5] - 2025-07-23
 
 ### Changed